+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 3.3.0
 * Added `PolicyMachine#all_operations_for_user_or_attr_and_objs_or_attrs` for finding a map of all operations
   between given objects and a given user.
+* Added `PM::Operation.prohibition?` method which takes a unique identifier and returns a boolean indicating
+  whether it represents a prohibition.
 
 ## 3.2.0
 * Added `PolicyMachine#accessible_objects_for_operations` for finding a map of accessible